+++ tests/test_cli.py
@@ -1,8 +1,9 @@
 """Test PRP cli functions."""
 
+import json
 from click.testing import CliRunner
 
-from prp.cli import create_output, print_schema
+from prp.cli import create_bonsai_input, create_cdm_input, print_schema
 
 
 def test_create_output_saureus(
@@ -26,7 +27,7 @@ def test_create_output_saureus(
     runner = CliRunner()
     with runner.isolated_filesystem():
         result = runner.invoke(
-            create_output,
+            create_bonsai_input,
             [
                 "-i",
                 sample_id,
@@ -50,6 +51,7 @@ def test_create_output_saureus(
                 saureus_mlst_path,
                 "--cgmlst",
                 saureus_chewbbaca_path,
+                "--output",
                 output_file,
             ],
         )
@@ -77,7 +79,7 @@ def test_create_output_ecoli(
     runner = CliRunner()
     with runner.isolated_filesystem():
         result = runner.invoke(
-            create_output,
+            create_bonsai_input,
             [
                 "-i",
                 sample_id,
@@ -101,15 +103,36 @@ def test_create_output_ecoli(
                 ecoli_mlst_path,
                 "--cgmlst",
                 ecoli_chewbbaca_path,
+                "--output",
                 output_file,
             ],
         )
         assert result.exit_code == 0
 
 
-def test_print_schema_cmd():
-    """Test print schema command."""
+def test_cdm_input_cmd(ecoli_quast_path, ecoli_bwa_path, ecoli_chewbbaca_path, ecoli_cdm_input):
+    """Test command for creating CDM input."""
     runner = CliRunner()
     with runner.isolated_filesystem():
-        result = runner.invoke(print_schema)
+        output_fname = "test_ouptut"
+        result = runner.invoke(
+            create_cdm_input,
+            [
+                "--quast",
+                ecoli_quast_path,
+                "--quality",
+                ecoli_bwa_path,
+                "--cgmlst",
+                ecoli_chewbbaca_path,
+                "--output",
+                output_fname,
+            ],
+        )
+
+        # test successful execution of command
         assert result.exit_code == 0
+
+        # test correct output format
+        with open(output_fname) as inpt:
+            cmd_output = json.load(inpt)
+            assert cmd_output == ecoli_cdm_input